requirements.txt
accelerate==1.4.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
blinker==1.9.0
boto3==1.37.9
botocore==1.37.9
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diffusers==0.32.2
Django==5.1.6
djangorestframework==3.15.2
filelock==3.13.1
fire==0.7.0
Flask==3.1.0
fsspec==2024.6.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psutil==7.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.4
safetensors==0.5.3
setuptools==70.2.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.21.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
